Make name required in axes
