+++ docs/changelog.md
@@ -11,6 +11,10 @@
 - Various optimisations have been made to reduce the number of file requests when creating a model class from an existing [zarr.Group][].
 - [ome_zarr_models.open_ome_zarr][] now includes the name of the group it was trying to validate alongside the validation error message.
 
+## Bug fixes
+
+- Axes metadata for version 0.4 and 0.5 has been fixed to require the `name` field.
+
 ## 1.1
 
 ### New Features
